feature: restructure files to presentational or smart components
