--- client/src/actions/fetchBusinesses.js
@@ -1,5 +1,0 @@
-// export function fetchBusinesses() {
-//   return (dispatch) => {
-//     axios.get('api/v1/businesses');
-//   };
-// }
