+++ client/src/components/businesses/smart/BusinessProfile.jsx
@@ -3,11 +3,11 @@ import ReactStars from 'react-stars';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { addFlashMessage } from '../../actions/flashMessages';
-import ReviewsCard from './ReviewsCards.jsx';
+import { addFlashMessage } from '../../../actions/flashMessages';
+import ReviewsCard from '../presentational/ReviewsCards.jsx';
 // import { title, description, category, location } from './RegisterBusiness.jsx';
-import { getReviewsAction, postReviewAction } from '../../actions/reviewsActions';
-import { getOneBusinessAction, deleteBusinessAction } from '../../actions/businessActions';
+import { getReviewsAction, postReviewAction } from '../../../actions/reviewsActions';
+import { getOneBusinessAction, deleteBusinessAction } from '../../../actions/businessActions';
 
 /**
    * @description -  Description of a given business profile
