+++ client/src/components/flash/FlashMessagesList.jsx
@@ -1,15 +1,23 @@
 import React from 'react';
-import { PropTypes } from 'prop-types';
-import FlashMessage from './FlashMessage';
 import { connect } from 'react-redux';
+import { PropTypes } from 'prop-types';
+import FlashMessage from './FlashMessage.jsx';
 import { deleteFlashMessage } from '../../actions/flashMessages';
 /**
  * @class FlashMessage
 */
 class FlashMessagesList extends React.Component {
+  /**
+ * @description Class component for Sogn up form
+ * @param { object } message
+ * @return { object } object
+*/
   render() {
     const messages = this.props.messages.map(message =>
-            <FlashMessage Key={message.id} message={message} deleteFlashMessage={this.props.deleteFlashMessage} />);
+           <FlashMessage
+            Key={message.id}
+            message={message} deleteFlashMessage={this.props.deleteFlashMessage}
+            />);
     return (
             <div>
                 {messages}
@@ -24,7 +32,11 @@ FlashMessagesList.propTypes = {
   deleteFlashMessage: PropTypes.func.isRequired
 };
 
-
+/**
+ * @description Class component for Sogn up form
+ * @param { object } state
+ * @return { object } object
+*/
 function mapStateToProps(state) {
   return {
     messages: state.flashMessages
