+++ client/src/components/home/smart/NavBar.jsx
@@ -2,7 +2,7 @@ import { PropTypes } from 'prop-types';
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { logout } from '../../actions/authActions';
+import { logout } from '../../../actions/authActions';
 
 /**
    * @description - class component for the navvbar
